python-server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
email-validator==2.2.0
pymongo==4.13.2
passlib==1.7.4
httpx==0.28.1
certifi==2026.2.25
PyJWT==2.10.1
